GOT THE MULITHREADED WORKING SUCKAAS
